[hotfix] Move TaskManagerSlot to o.a.f.runtime.resourcemanager.slotmanager
